리팩토링: AITConvertCore WebGL 빌드 결과 진단 로직을 AITWebGLBuildDiagnostics로 분리 (#36)
god-class Editor/AITConvertCore.cs(1,416행)에서 빌드 결과 해석 헬퍼를 별도 클래스로 추출:
- MapBuildResultToExportError (Unity BuildResult → AIT 에러 코드 매핑)
- BuildFailureSummary (에러/경고 캡 분리 실패 요약 문자열)
- ReadEditorLogTail (무진단 실패 대비 에디터 로그 꼬리)
- FailureSummaryMaxErrors/MaxWarnings 상수

동작 변경 없음(behavior-preserving): 본문 그대로 이동하고 내부 호출부 2곳과
EditMode 테스트 참조(BuildFailureSummaryTests/BuildResultMappingTests)를 새 클래스로 갱신.
공개 계약인 AITExportError 열거형과 GetError* 메시지 변환은 AITConvertCore에 유지.

이름이 비슷한 기존 AppsInToss.Editor.AITBuildDiagnostics(Sentry 진단 버퍼)와의 혼동을
피하려 WebGL 접두사로 명명하고 클래스 doc에 차이를 명시.

AITConvertCore.cs: 1,416 → 1,268행 (-148)
